Add client UI

+++ client/Assets/Scripts/SystemState.cs
@@ -1,6 +1,6 @@
 ﻿[System.Serializable]
 public struct SystemState
 {
-    public int timestamp;
+    public float timestamp;
     public float solarPanelVoltage;
 }
